Revert "Move EnvironmentSettingNames to infrastructure"
This reverts commit 60454153c6389a9d03273109389c7db874d9b1c7.
