+++ source/TimeSeries/TimeSeriesBundleIngestor/ServiceCollectionExtensions.cs
@@ -17,7 +17,6 @@ using Energinet.DataHub.Core.App.Common.Abstractions.Security;
 using Energinet.DataHub.Core.App.Common.Identity;
 using Energinet.DataHub.Core.App.Common.Security;
 using Energinet.DataHub.Core.App.FunctionApp.Middleware;
-using Energinet.DataHub.TimeSeries.Infrastructure;
 using Energinet.DataHub.TimeSeries.Infrastructure.Registration;
 using Microsoft.Extensions.DependencyInjection;
